RandomID as filename for uploaded image and error handling.
The error handling is a catch-all without trying to correct
anything - only stating that something went wrong, instead of
crashing the program :)

The RandomID is inspired by the Go implementation that can be
found in `thread-art-archive`. Even though the ID is 12 characters
long, there could occur a collision, where the correct way is
maybe to tell the user to try again, or automatically generate
another ID until not colliding.