Queue handling.
A temporary file is created at the start of the request.
If there are more than QUEUE_LIMIT files in the queue, then
the file is removed. Otherwise we try to handle the request and
add it to the queue. However if an error occurs later, then the
request will still be in the queue: which can cause a problem
later. The way this is handled might also cause a problem, because
act - though not yet implemented - will try to handle the next
item in the queue instantly (if there is an available worker).

It might be mitigated by adding a byte to the queue file signalling
that the image has been successfully saved.

If an error occurs it should remove the temporary file.